uploads 07/11/25

+++ app/Http/Controllers/LoaiSanPhamController.php
@@ -5,12 +5,14 @@ namespace App\Http\Controllers;
 use App\Models\LoaiSanPham;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
+use Illuminate\Validation\Rule;
 
 class LoaiSanPhamController extends Controller
 {
+    // Lấy danh sách
     public function getDanhSach()
     {
-        $loaisanpham = LoaiSanPham::all();
+        $loaisanpham = LoaiSanPham::orderBy('tenloai', 'asc')->get();
         return view('admin.loaisanpham.danhsach', compact('loaisanpham'));
     }
 
@@ -21,47 +23,46 @@ class LoaiSanPhamController extends Controller
 
     public function postThem(Request $request)
     {
-        //luật
         $request->validate([
             'tenloai' => ['required', 'string', 'max:255', 'unique:loaisanpham,tenloai'],
         ]);
 
         $orm = new LoaiSanPham();
-        $orm->tenloai = $request->tenloai;
-        $orm->tenloai_slug = Str::slug($request->tenloai, '-');
+        $orm->tenloai = Str::title($request->tenloai);
+        $orm->tenloai_slug = Str::slug($orm->tenloai, '-');
         $orm->save();
 
         return redirect()->route('admin.loaisanpham')->with('success', 'Thêm loại sản phẩm thành công!');
-
     }
 
     public function getSua($id)
     {
-        $loaisanpham = LoaiSanPham::find($id);
+        $loaisanpham = LoaiSanPham::findOrFail($id);
+
         return view('admin.loaisanpham.sua', compact('loaisanpham'));
     }
 
     public function postSua(Request $request, $id)
     {
-        //kiểm tra
-        $request = validate([
-            'tenloai' => ['required', 'string', 'max:255', 'unique:loaisanpham,tenloai' . $id],
+        $request->validate([
+            'tenloai' => ['required', 'string', 'max:255', Rule::unique('loaisanpham', 'tenloai')->ignore($id)],
         ]);
 
-        $orm = LoaiSanPham::find($id);
-        $orm->tenloai = $request->tenloai;
-        $orm->tenloai_slug = Str::slug($request->tenloai, '-');
+        $orm = LoaiSanPham::findOrFail($id);
+        $orm->tenloai = Str::title($request->tenloai);
+        $orm->tenloai_slug = Str::slug($orm->tenloai, '-');
         $orm->save();
 
-        return redirect()->route('admin.loaisanpham');
+        return redirect()->route('admin.loaisanpham')->with('success', 'Cập nhật loại sản phẩm thành công!');
     }
 
     public function getXoa($id)
     {
         $orm = LoaiSanPham::find($id);
-        $orm->delete();
-
-        return redirect()->route('admin.loaisanpham');
+        if ($orm) {
+            $orm->delete();
+            return redirect()->route('admin.loaisanpham')->with('success', 'Xóa thành công');
+        }
+        return redirect()->route('admin.loaisanpham')->with('error', 'Lỗi khi xóa');
     }
-
 }
